+++ libs/ngx-editorjs-demo/src/lib/plugins/image/image.plugin.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { ToolSettings } from '@editorjs/editorjs';
 import Image from '@editorjs/image';
-import { EditorJSPlugin } from '@tinynodes/ngx-editorjs';
+import { BasePlugin } from '@tinynodes/ngx-editorjs';
 
 @Injectable()
-export class PluginImage extends EditorJSPlugin {
-  static plugin(): ToolSettings {
+export class PluginImage implements BasePlugin {
+  plugin(): ToolSettings {
     return Image;
   }
 }