+++ libs/ngx-editorjs-demo/src/lib/plugins/marker/marker.plugin.ts
@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
 import { ToolSettings } from '@editorjs/editorjs';
 import Marker from '@editorjs/marker';
-import { EditorJSPlugin } from '@tinynodes/ngx-editorjs';
+import { BasePlugin } from '@tinynodes/ngx-editorjs';
 
 @Injectable()
-export class PluginMarker extends EditorJSPlugin {
-  static plugin(): ToolSettings {
-    return {
-      class: Marker,
-      shortcut: 'CMD+SHIFT+M'
-    };
+export class PluginMarker implements BasePlugin {
+  plugin(): ToolSettings {
+    return Marker;
+  }
+
+  shortcut() {
+    return 'CMD+SHIFT+M';
   }
 }