+++ libs/ngx-editorjs/src/lib/plugins/paragraph/paragraph.module.ts
@@ -1,0 +1,7 @@
+import { NgModule } from '@angular/core';
+import { PluginParagraph } from './paragraph.plugin';
+
+@NgModule({
+  providers: [PluginParagraph]
+})
+export class PluginParagraphModule {}