+++ libs/ngx-editorjs/src/lib/plugins/paragraph/paragraph.plugin.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import Header from '@editorjs/header';
+import Paragraph from '@editorjs/paragraph';
 import { ToolSettings } from '@editorjs/editorjs';
 import { BasePlugin } from '../../types/plugins';
 
 @Injectable()
-export class PluginHeader implements BasePlugin {
+export class PluginParagraph implements BasePlugin {
   public plugin(): ToolSettings {
-    return Header;
+    return Paragraph;
   }
 }